move theme.js file back
