+++ client/src/Components/Pages/Login/Login.style.js
@@ -1,7 +1,7 @@
 import { Link } from "react-router-dom";
 import styled from "styled-components";
 
-import { colors } from "./../../../theme";
+import { colors } from "./../../../Components/theme";
 
 export const LoginWrapper = styled.div.attrs({ className: "login" })`
   width: 100%;
